Move ProcessingStatus to model
