+++ frontend/src/main/scala/hmda/api/http/HttpApi.scala
@@ -4,26 +4,32 @@ import java.net.InetAddress
 import java.time.Instant
 
 import akka.Done
-import akka.actor.ActorSystem
+import akka.actor.{ ActorSystem, PoisonPill }
 import akka.event.LoggingAdapter
 import akka.http.scaladsl.marshalling.ToResponseMarshallable
-import akka.http.scaladsl.model.{ HttpResponse, Multipart, StatusCodes }
+import akka.http.scaladsl.model._
 import akka.http.scaladsl.model.Multipart.BodyPart
 import akka.http.scaladsl.server.Directives._
 import akka.stream.ActorMaterializer
 import akka.stream.scaladsl.{ Framing, Sink }
-import akka.util.ByteString
+import akka.pattern.ask
+import akka.util.{ ByteString, Timeout }
 import hmda.api.model.Status
 import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport._
+import hmda.api.model.processing.ProcessingStatusActor
+import hmda.api.model.processing.ProcessingStatusActor.GetProcessingStatus
 import hmda.api.protocol.HmdaApiProtocol
+import hmda.api.protocol.processing.ProcessingStatusProtocol
+import hmda.model.messages.ProcessingStatus
 import hmda.persistence.HmdaFileRaw
 import hmda.persistence.HmdaFileRaw.{ AddLine, CompleteUpload, Shutdown }
 import spray.json._
 
+import scala.concurrent.duration._
 import scala.concurrent.Future
 import scala.util.{ Failure, Success }
 
-trait HttpApi extends HmdaApiProtocol {
+trait HttpApi extends HmdaApiProtocol with ProcessingStatusProtocol {
 
   implicit val system: ActorSystem
   implicit val materializer: ActorMaterializer
@@ -77,9 +83,18 @@ trait HttpApi extends HmdaApiProtocol {
     }
   }
 
-  val uploadedStatsPath = path("uploaded" / Segment) { id =>
-    complete("ok")
-  }
+//  val processingStatusPath = path("status" / Segment) { id =>
+//    //TODO: make timeout configurable?
+//    implicit val timeout = Timeout(5.seconds)
+//    implicit val ec = system.dispatcher
+//    get {
+//      val processingActor = system.actorOf(ProcessingStatusActor.props(id))
+//      (processingActor ? GetProcessingStatus).mapTo[List[ProcessingStatus]]
+//        .map(result => complete(ToResponseMarshallable(result)))
+//        .recover { case ex => complete(HttpResponse(StatusCodes.InternalServerError)) }
+//    }
+//
+//  }
 
-  val routes = rootPath ~ uploadPath ~ uploadedStatsPath
+  val routes = rootPath ~ uploadPath //~ processingStatusPath
 }
