--- frontend/src/main/scala/hmda/api/model/processing/ProcessingStatus.scala
@@ -1,3 +1,0 @@
-package hmda.api.model.processing
-
-case class ProcessingStatus(id: String, dateStarted: String, rowsUploaded: Int)
