+++ model/src/main/scala/hmda/model/messages/ProcessingStatus.scala
@@ -1,0 +1,5 @@
+package hmda.model.messages
+
+import java.time.Instant
+
+case class ProcessingStatus(id: String = "", dateStarted: String = Instant.now.toString, rowsUploaded: Int = 0)
